+++ build.gradle
@@ -51,6 +51,7 @@ dependencies {
 
     implementation 'com.google.api-client:google-api-client:1.32.1'
     implementation 'com.google.http-client:google-http-client-jackson2:1.39.2'
+    implementation 'com.squareup.okhttp3:okhttp:4.10.0'
 
     implementation "org.springframework.boot:spring-boot-starter-webflux"
