+++ src/main/java/com/m9d/sroom/util/youtube/WebClientService.java
@@ -6,11 +6,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.util.DefaultUriBuilderFactory;
 import org.springframework.web.util.UriComponentsBuilder;
 import reactor.core.publisher.Mono;
 
-@Service
+//@Service
 @Slf4j
 @RequiredArgsConstructor
 public class WebClientService implements YoutubeApi {
