+++ src/main/resources/application-local.yml
@@ -23,9 +23,6 @@ spring:
 
 youtube:
   base-url: https://www.googleapis.com/youtube/v3
-  schema: https
-  host: www.googleapis.com
-  path: youtube/v3
 
 #Devtool
 devtools:
