+++ src/test/resources/application.yml
@@ -21,7 +21,7 @@ spring:
     password:
 
 youtube:
-  base-url: https://www.googleapis.com/youtube/v3/
+  base-url: https://www.googleapis.com/youtube/v3
 
 #Devtool
 devtools:
